tidy up intercom code by making it the 5th part of exhibitor
