+++ src/comp-site/tsconfig.json
@@ -3,7 +3,8 @@
   "extends": "../../tsconfig.json",
   "include": [
     "./**/*",
-    "../common/**/*"
+    "../common/**/*",
+    "../intercom/**/*"
   ],
   "exclude": [
     "./vue/**/*"
