+++ src/intercom/message/index.ts
@@ -1,0 +1,14 @@
+import { IntercomIdentityType } from '../types'
+import { IntercomMessage, IntercomMessageOptions } from './types'
+
+export const createIntercomMessage = (
+  from: IntercomIdentityType,
+  options: IntercomMessageOptions,
+): IntercomMessage => ({
+  from,
+  ...options,
+})
+
+export const sendIntercomMessage = (ws: WebSocket, msg: IntercomMessage) => {
+  ws.send(JSON.stringify(msg))
+}
