+++ src/site/server/tsconfig.json
@@ -4,7 +4,8 @@
   "include": [
     "./**/*",
     "../common/**/*",
-    "../../common/**/*"
+    "../../common/**/*",
+    "../../intercom/**/*"
   ],
   "compilerOptions": {
     "noEmit": true
